Enviando exercicio praticado a031 componentes
