+++ src/componentes/Usuario.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+export function Usuario() { //criando e exportando
+
+    return (
+
+        <div>
+            <img />
+            <p>Ronildo</p>
+        </div>
+
+
+    )
+
+}
+
